docs: Put demo code in their own files
