+++ website/package.json
@@ -19,6 +19,7 @@
     "@types/react-dom": "^17.0.0",
     "@types/react-helmet": "^6.1.5",
     "@types/react-router-dom": "^5.3.3",
+    "raw-loader": "^4.0.2",
     "serve": "14.2.0",
     "webpack": "^5.74.0"
   },
