+++ website/src/components/Demo/code/simple/fetch/Demo.tsx
@@ -1,0 +1,7 @@
+import { getTodo } from './api';
+
+function TodoDetail({ id }: { id: number }) {
+  const todo = useSuspense(getTodo, id);
+  return <div>{todo.title}</div>;
+}
+render(<TodoDetail id={1} />);
