+++ website/src/components/Demo/code/simple/graphql/api.ts
@@ -1,0 +1,11 @@
+import { GQLEndpoint } from '@rest-hooks/graphql';
+
+const gql = new GQLEndpoint('/');
+export const getTodo = gql.query(`query GetTodo($id: ID!) {
+  todo(id: $id) {
+    id
+    title
+    completed
+  }
+}
+`);
