+++ website/src/components/Demo/code/simple/rest/Demo.tsx
@@ -1,0 +1,7 @@
+import { TodoResource } from './api';
+
+function TodoDetail({ id }: { id: number }) {
+  const todo = useSuspense(TodoResource.get, { id });
+  return <div>{todo.title}</div>;
+}
+render(<TodoDetail id={1} />);
