+++ website/src/components/Demo/code/todo-app/graphql/TodoItem.tsx
@@ -1,0 +1,21 @@
+import { TodoResource, type Todo } from './api';
+
+export default function TodoItem({ todo }: { todo: Todo }) {
+  const controller = useController();
+  return (
+    <div>
+      <label>
+        <input
+          type="checkbox"
+          checked={todo.completed}
+          onChange={e =>
+            controller.fetch(TodoResource.update, {
+              todo: { id: todo.id, completed: e.currentTarget.checked },
+            })
+          }
+        />
+        {todo.completed ? <strike>{todo.title}</strike> : todo.title}
+      </label>
+    </div>
+  );
+}
