+++ website/src/components/Demo/code/todo-app/graphql/api.ts
@@ -1,0 +1,55 @@
+import { GQLEndpoint, GQLEntity } from '@rest-hooks/graphql';
+
+const gql = new GQLEndpoint('/');
+
+export class Todo extends GQLEntity {
+  readonly title: string = '';
+  readonly completed: boolean = false;
+  readonly userId: number = 0;
+}
+
+export const TodoResource = {
+  get: gql.query(
+    `query GetTodo($id: ID!) {
+    todo(id: $id) {
+      id
+      title
+      completed
+      userId
+    }
+  }
+`,
+    { todo: Todo },
+  ),
+  getList: gql.query(
+    `query GetTodos {
+    todo {
+      id
+      title
+      completed
+      userId
+    }
+  }
+`,
+    { todos: [Todo] },
+  ),
+  update: gql.mutation(
+    `mutation UpdateTodo($todo: Todo!) {
+    updateTodo(todo: $todo) {
+      id
+      title
+      completed
+    }
+  }`,
+    { updateTodo: Todo },
+  ),
+  queryRemaining: new Query(
+    new schema.All(Todo),
+    (entries, { userId } = {}) => {
+      if (userId !== undefined)
+        return entries.filter(todo => todo.userId === userId && !todo.completed)
+          .length;
+      return entries.filter(todo => !todo.completed).length;
+    },
+  ),
+};
