+++ website/src/components/Demo/code/todo-app/index.ts
@@ -1,0 +1,4 @@
+import graphqlDemo from './graphql';
+import restDemo from './rest';
+
+export default [restDemo, graphqlDemo];
